*add connect and socket timeout to httpclient inside middlewares; *remove some unused codes;
