+++ heimdall-gateway/src/test/java/br/com/conductor/heimdall/gateway/filter/HeimdallDecorationFilterTest.java
@@ -1,24 +1,16 @@
 
 package br.com.conductor.heimdall.gateway.filter;
 
-import br.com.conductor.heimdall.core.enums.HttpMethod;
-import br.com.conductor.heimdall.gateway.router.Credential;
-import br.com.conductor.heimdall.gateway.router.CredentialRepository;
-import br.com.conductor.heimdall.gateway.router.EnvironmentInfo;
-import br.com.conductor.heimdall.gateway.router.EnvironmentInfoRepository;
-import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
-import br.com.conductor.heimdall.gateway.util.RequestHelper;
-import br.com.conductor.heimdall.gateway.zuul.route.HeimdallRoute;
-import br.com.conductor.heimdall.gateway.zuul.route.ProxyRouteLocator;
-import com.google.common.collect.Sets;
-import com.netflix.zuul.context.RequestContext;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.REQUEST_URI_KEY;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -36,14 +28,17 @@ import org.springframework.http.HttpStatus;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicReference;
 import com.netflix.zuul.context.RequestContext;
 
-import static org.junit.Assert.*;
-import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.REQUEST_URI_KEY;
+import br.com.conductor.heimdall.core.enums.HttpMethod;
+import br.com.conductor.heimdall.gateway.router.Credential;
+import br.com.conductor.heimdall.gateway.router.CredentialRepository;
+import br.com.conductor.heimdall.gateway.router.EnvironmentInfo;
+import br.com.conductor.heimdall.gateway.router.EnvironmentInfoRepository;
+import br.com.conductor.heimdall.gateway.trace.TraceContextHolder;
+import br.com.conductor.heimdall.gateway.util.RequestHelper;
+import br.com.conductor.heimdall.gateway.zuul.route.HeimdallRoute;
+import br.com.conductor.heimdall.gateway.zuul.route.ProxyRouteLocator;
 
 @RunWith(MockitoJUnitRunner.class)
 public class HeimdallDecorationFilterTest {
